requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-python==9.5.0
openpyxl==3.1.5
PyMySQL==1.1.2
waitress==3.0.2
Werkzeug==3.1.5
xlsxwriter==3.2.9